ecu_tuner_project_refined_package/requirements.txt
Flask
numpy
pyserial
pyttsx3
openai
python-dotenv
requests
SpeechRecognition